refactor: Adjust relative import paths for API and type imports in admin hooks.
